workshops/hybrid-cloud-workshop-one-cluster-linking/pyproject.toml
[tool.poetry]
name = "hybrid-cloud-workshop-one-cluster-linking"
version = "0.1.0"
description = ""
authors = ["Adi Wabisabi <awabisabi@confluent.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0.1"
boto3 = "^1.34.36"
google = "^3.0.0"
google-api-python-client = "^2.116.0"
azure-cli = "^2.57.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
